+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # api
 
+## 10x.19.4
+- Extend wiki details with profile data
+
+## 10x.19.3
+- Set specific Helm and Helmfile versions
+
+## 10x.19.2
+- Add number of triples count to the wiki daily metrics
+
 ## 10x.19.1
 - no changes, new tag to unclog CI